requirements.txt
accelerate==1.10.1
acres==0.5.0
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appnope @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_46iquy4y0r/croot/appnope_1750774774351/work
asttokens @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_11i8cnwuxs/croot/asttokens_1743630449718/work
attrs==25.3.0
Authlib==1.6.3
backoff==2.2.1
beautifulsoup4==4.13.5
bert-score==0.3.13
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
circuitbreaker==2.1.3
click==8.2.1
cloudpickle @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_04keaza5jy/croot/cloudpickle_1751309451361/work
comm @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/comm_1728587617427/work
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==44.0.3
cycler==0.12.1
cyclopts==3.24.0
debugpy @ file:///opt/miniconda3/conda-bld/debugpy_1757067177426/work
decorator @ file:///opt/miniconda3/conda-bld/decorator_1757341238163/work
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docling==2.52.0
docling-core==2.48.1
docling-ibm-models==3.9.1
docling-parse==4.4.0
docstring_parser==0.17.0
docutils==0.22
easyocr==1.7.2
email-validator==2.3.0
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
executing @ file:///opt/miniconda3/conda-bld/executing_1757061235986/work
faiss-cpu==1.12.0
Faker==37.6.0
fastapi==0.116.2
fastmcp==2.12.3
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.0
frontend==0.0.3
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-auth==2.40.3
google-genai==1.36.0
groq==0.31.1
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
ipykernel @ file:///opt/miniconda3/conda-bld/ipykernel_1757583370966/work
ipython @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_2evnd74hdv/croot/ipython_1734548057210/work
isodate==0.7.2
itsdangerous==2.2.0
jedi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_38ctoinnl0/croot/jedi_1733987402850/work
Jinja2==3.1.6
jiter==0.11.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1756321760188/work
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client @ file:///Users/builder/cbouss/buildout/croot/jupyter_client_1739483928377/work
jupyter_core @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_044bmfoejx/croot/jupyter_core_1751991384131/work
kiwisolver==1.4.9
latex2mathml==3.78.1
lazy-object-proxy==1.12.0
lazy_loader==0.4
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline @ file:///Users/builder/cbouss/buildout/croot/matplotlib-inline_1732916182598/work
mcp==1.14.0
mdurl==0.1.2
mem0ai==0.1.117
more-itertools==10.8.0
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
narwhals==2.5.0
nest_asyncio @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/nest-asyncio_1728588080290/work
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numpy==2.2.6
oci==2.160.1
openai==1.99.9
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b31rpdx1of/croot/packaging_1753775365574/work
pandas==2.3.2
parse==1.20.2
parso @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8824a1w4md/croot/parso_1733963320105/work
pathable==0.4.4
pathlib==1.0.1
pexpect @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_93wbojua0f/croot/pexpect_1750072609163/work
pillow==11.2.1
pip-tools==7.5.0
platformdirs @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4euggz8v1n/croot/platformdirs_1744273055841/work
pluggy==1.6.0
polyfactory==2.22.2
portalocker==3.2.0
postgrest==1.1.1
posthog==6.7.4
prompt_toolkit @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/prompt-toolkit_1728614904988/work
protobuf==5.29.5
prov==2.1.1
psutil @ file:///opt/miniconda3/conda-bld/psutil_1757104544499/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///opt/miniconda3/conda-bld/pure_eval_1757067068750/work
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
Pygments @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f0f10r98sf/croot/pygments_1744664126614/work
PyJWT==2.10.1
pylatexenc==2.10
pyOpenSSL==24.3.0
pyparsing==3.2.4
pypdfium2==4.30.0
pyperclip==1.9.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/python-dateutil_1728585579455/work
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq @ file:///opt/miniconda3/conda-bld/pyzmq_1758103939568/work
qdrant-client==1.15.1
rdflib==7.1.4
realtime==2.7.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1757406203051/work/dist/scikit_learn-1.7.2-cp313-cp313-macosx_11_0_arm64.whl#sha256=2c26ddc964d74714c5966793740443210e988fc9c6f99317d4e853bc87a82ede
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1757681124245/work/dist/scipy-1.16.2-cp313-cp313-macosx_11_0_arm64.whl#sha256=e45586ddb12f198a75c10c57ce70bf029a9147d9f620454ccc3cb1b962262901
selenium==4.35.0
semchunk==2.2.2
sentence-transformers==5.1.0
setuptools==78.1.1
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03myqm7p6o/croot/six_1744271511946/work
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spyder-kernels @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f1f2blkxlb/croot/spyder-kernels_1748642516744/work
SQLAlchemy==2.0.43
sse-starlette==3.0.2
stack_data @ file:///opt/miniconda3/conda-bld/stack_data_1757067040463/work
starlette==0.48.0
storage3==0.12.1
streamlit==1.49.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tifffile==2025.9.9
tokenizers==0.22.0
toml==0.10.2
tools==1.0.5
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/traitlets_1728585482264/work
traits==7.0.2
transformers==4.56.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_8el7txr34i/croot/wcwidth_1750352901095/work
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wheel==0.45.1
wsproto==1.2.0
wurlitzer @ file:///opt/miniconda3/conda-bld/wurlitzer_1756979036322/work
xlsxwriter==3.2.8
